requirements.txt
PyQt5>=5.15.0
Pillow>=10.0.0
requests>=2.31.0
numpy>=1.24.0
openpyxl>=3.1.0
